# docs/openapi/product-service-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "商品服务 API",
    "version": "1.0.0",
    "description": "product-service 接口定义文档。"
  },
  "servers": [
    {
      "url": "http://localhost:8080/api/product",
      "description": "本地网关(商品服务)"
    },
    {
      "url": "http://localhost:8082",
      "description": "本地商品服务"
    }
  ],
  "tags": [
    {
      "name": "商品管理",
      "description": "商品管理相关接口"
    }
  ],
  "paths": {
    "/products": {
      "get": {
        "tags": [
          "商品管理"
        ],
        "summary": "查询商品列表",
        "operationId": "listProducts",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseProductList"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "商品管理"
        ],
        "summary": "创建商品（需权限点 products:create）",
        "operationId": "createProduct",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductCreateRequest"
              },
              "example": {
                "productName": "Mechanical Keyboard",
                "categoryId": 5,
                "price": 199.00,
                "stock": 100,
                "status": 1,
                "remark": "RGB mechanical keyboard"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseProduct"
                }
              }
            }
          }
        }
      }
    },
    "/products/{id}": {
      "get": {
        "tags": [
          "商品管理"
        ],
        "summary": "查询商品详情",
        "operationId": "getProductById",
        "parameters": [
          {
            "$ref": "#/components/parameters/ProductId"
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseProduct"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "商品管理"
        ],
        "summary": "更新商品（需权限点 products:update）",
        "operationId": "updateProduct",
        "parameters": [
          {
            "$ref": "#/components/parameters/ProductId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductUpdateRequest"
              },
              "example": {
                "productName": "Mechanical Keyboard Pro",
                "categoryId": 5,
                "price": 219.00,
                "stock": 80,
                "status": 1,
                "remark": "Updated version"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseProduct"
                }
              }
            }
          }
        }
      }
    },
    "/products/{id}/status": {
      "patch": {
        "tags": [
          "商品管理"
        ],
        "summary": "修改商品状态（需权限点 products:status）",
        "operationId": "changeProductStatus",
        "parameters": [
          {
            "$ref": "#/components/parameters/ProductId"
          },
          {
            "name": "status",
            "in": "query",
            "required": true,
            "description": "1 表示上架，0 表示下架",
            "schema": {
              "type": "integer",
              "enum": [
                0,
                1
              ]
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseProduct"
                }
              }
            }
          }
        }
      }
    },
    "/products/{id}/stock": {
      "post": {
        "tags": [
          "商品管理"
        ],
        "summary": "调整商品库存（需权限点 products:stock）",
        "operationId": "adjustProductStock",
        "parameters": [
          {
            "$ref": "#/components/parameters/ProductId"
          },
          {
            "name": "delta",
            "in": "query",
            "required": true,
            "description": "库存增量，负数表示扣减库存，正数表示回补库存",
            "schema": {
              "type": "integer"
            },
            "example": -2
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseProduct"
                }
              }
            }
          }
        }
      }
    },
    "/products/internal/{id}/stock": {
      "post": {
        "tags": [
          "商品管理"
        ],
        "summary": "内部调整库存（服务间调用，不设权限门槛）",
        "description": "仅供 order-service 经 Feign 调用，不经网关、无用户角色上下文。生产环境应叠加内网隔离或内部调用凭证。",
        "operationId": "adjustProductStockInternal",
        "parameters": [
          {
            "$ref": "#/components/parameters/ProductId"
          },
          {
            "name": "delta",
            "in": "query",
            "required": true,
            "description": "库存增量，负数表示扣减库存，正数表示回补库存",
            "schema": {
              "type": "integer"
            },
            "example": -2
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseProduct"
                }
              }
            }
          }
        }
      }
    },
    "/products/page": {
      "get": {
        "tags": [
          "商品管理"
        ],
        "summary": "分页查询商品列表",
        "operationId": "pageProducts",
        "parameters": [
          {
            "name": "pageNum",
            "in": "query",
            "required": false,
            "description": "页码，默认 1",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "example": 1
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "description": "每页条数，默认 10",
            "schema": {
              "type": "integer",
              "default": 10
            },
            "example": 10
          },
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "description": "模糊搜索关键字（匹配 productName/remark）",
            "schema": {
              "type": "string"
            },
            "example": "Keyboard"
          },
          {
            "name": "categoryId",
            "in": "query",
            "required": false,
            "description": "按分类 ID 筛选，命中索引 idx_products_category。不传则不限分类。注意：与 keyword 同时传时为 AND 关系",
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 1
            },
            "example": 5
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseProductPage"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "parameters": {
      "ProductId": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "商品 ID",
        "schema": {
          "type": "integer",
          "format": "int64"
        },
        "example": 1
      }
    },
    "schemas": {
      "ApiResponseProduct": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 200
          },
          "message": {
            "type": "string",
            "example": "success"
          },
          "data": {
            "$ref": "#/components/schemas/Product"
          }
        }
      },
      "ApiResponseProductList": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 200
          },
          "message": {
            "type": "string",
            "example": "success"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Product"
            }
          }
        }
      },
      "ApiResponseProductPage": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 200
          },
          "message": {
            "type": "string",
            "example": "success"
          },
          "data": {
            "$ref": "#/components/schemas/ProductPage"
          }
        }
      },
      "ProductPage": {
        "type": "object",
        "properties": {
          "records": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Product"
            }
          },
          "total": {
            "type": "integer",
            "format": "int64",
            "description": "总记录数",
            "example": 50
          },
          "pages": {
            "type": "integer",
            "format": "int64",
            "description": "总页数",
            "example": 5
          },
          "current": {
            "type": "integer",
            "description": "当前页码（从 1 开始）",
            "example": 1
          },
          "size": {
            "type": "integer",
            "description": "每页条数",
            "example": 10
          }
        }
      },
      "Product": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "productName": {
            "type": "string",
            "example": "Mechanical Keyboard"
          },
          "categoryId": {
            "type": "integer",
            "format": "int64",
            "description": "分类 ID，可为空（未分类商品为 null）",
            "example": 5
          },
          "price": {
            "type": "number",
            "format": "double",
            "example": 199.00
          },
          "stock": {
            "type": "integer",
            "example": 100
          },
          "status": {
            "type": "integer",
            "description": "1 表示上架，0 表示下架",
            "example": 1
          },
          "remark": {
            "type": "string",
            "example": "RGB mechanical keyboard"
          },
          "createTime": {
            "type": "string",
            "format": "date-time",
            "example": "2026-07-12T12:00:00"
          },
          "updateTime": {
            "type": "string",
            "format": "date-time",
            "example": "2026-07-12T12:30:00"
          }
        }
      },
      "ProductCreateRequest": {
        "type": "object",
        "required": [
          "productName",
          "price",
          "stock",
          "status"
        ],
        "properties": {
          "productName": {
            "type": "string"
          },
          "categoryId": {
            "type": "integer",
            "format": "int64",
            "minimum": 1,
            "description": "分类 ID，可选；不传表示不设置分类"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "stock": {
            "type": "integer"
          },
          "status": {
            "type": "integer",
            "enum": [
              0,
              1
            ]
          },
          "remark": {
            "type": "string"
          }
        }
      },
      "ProductUpdateRequest": {
        "type": "object",
        "required": [
          "productName",
          "price",
          "stock",
          "status"
        ],
        "properties": {
          "productName": {
            "type": "string"
          },
          "categoryId": {
            "type": "integer",
            "format": "int64",
            "minimum": 1,
            "description": "分类 ID，可选；不传表示不设置分类"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "stock": {
            "type": "integer"
          },
          "status": {
            "type": "integer",
            "enum": [
              0,
              1
            ]
          },
          "remark": {
            "type": "string"
          }
        }
      }
    }
  }
}